+++ clean_start.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 rm -f /home/vagrant/marketplace_db.db
-python marketplace.py
+python run.py
