+++ marketplace/routes/home.py
@@ -1,0 +1,9 @@
+from flask import Blueprint
+
+home = Blueprint('name', __name__)
+
+
+@home.route('/')
+def home_page():
+    """ Our home page. """
+    return "In progress"
