+++ run.py
@@ -1,0 +1,4 @@
+from marketplace import app
+
+if __name__ == '__main__':
+    app.run(host=app.config['HOST'])
